pyproject.toml
[project]
name = "secret-mission"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "html2text>=2025.4.15",
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
    "python-slugify>=8.0.4",
    "requests>=2.32.5",
]
